+++ scripts/gulp/tasks/bundle.ts
@@ -3,8 +3,9 @@ import babel from 'gulp-babel';
 import sourcemaps from 'gulp-sourcemaps';
 import del from 'del';
 
-import config from '../config';
 import { sh } from '../../utils';
+
+import config from '../config';
 
 const { paths } = config;
