+++ scripts/gulp/webpack/webpack.config.e2e.ts
@@ -1,7 +1,9 @@
 import CopyWebpackPlugin from 'copy-webpack-plugin';
 import ForkTsCheckerWebpackPlugin from 'fork-ts-checker-webpack-plugin';
 import webpack from 'webpack';
+
 import { getDefaultEnvironmentVars } from '../../monorepo';
+
 import config from '../config';
 
 const { paths } = config;
